backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
sendgrid
scikit-learn
joblib
python-jose
PyJWT
